pyproject.toml
[project]
name = "pg-plus"
version = "0.1.0"
description = "PG-Plus backend + scripts scaffold managed via uv"
requires-python = ">=3.11,<3.12"
dependencies = [
    "Django>=5.0,<6.0",
    "djangorestframework>=3.15,<3.16",
    "djangorestframework-simplejwt>=5.3,<6.0",
    "django-cors-headers>=4.3,<5.0",
    "mysqlclient>=2.2,<3.0",
    "celery>=5.3,<6.0",
    "redis>=5.0,<6.0",
    "django-celery-results>=2.5,<3.0",
    "django-celery-beat>=2.6,<3.0",
    "gunicorn>=21.2,<22.0",
    "uvicorn[standard]>=0.27,<0.28",
]

[dependency-groups]
dev = [
    "black>=23.12,<25.0",
    "isort>=5.13,<6.0",
    "flake8>=7.0,<8.0",
]
